add an option to turn off the sound file if it has loop settings Delete lcm and gcd calculations in BeatFraction class constructor fix TRACE log compile error

+++ libTest/BMSDecryptor.h
@@ -100,7 +100,7 @@ namespace bms {
 		/// </summary>
 		inline long long GetTimeUsingBeat(const BeatFraction& beat) {
 			BeatFraction subtract;
-			size_t index = static_cast<uint32_t>(mData.mListTimeSeg.size()) - 1;
+			uint32_t index = static_cast<uint32_t>(mData.mListTimeSeg.size()) - 1;
 			for (; index > 0; --index) {
 				const TimeSegment& t = mData.mListTimeSeg[index];
 				subtract = beat - t.mCurBeat;
